添加了HoloScore装饰品 添加了3D渲染相关

+++ RandomBuff/Core/Game/Settings/Conditions/AchievementCondition.cs
@@ -29,7 +29,7 @@ namespace RandomBuff.Core.Game.Settings.Conditions
 
         public override ConditionID ID => ConditionID.Achievement;
 
-        public override float Exp => 100;
+        public override int Exp => 100;
 
         public override ConditionState SetRandomParameter(SlugcatStats.Name name, float difficulty,
             List<Condition> sameConditions = null)
