+++ RandomBuff/Core/Game/Settings/Conditions/HuntCondition.cs
@@ -12,7 +12,7 @@ namespace RandomBuff.Core.Game.Settings.Conditions
     {
         public override ConditionID ID => ConditionID.Hunt;
 
-        public override float Exp => 0;
+        public override int Exp => 0;
 
         [JsonProperty]
         public CreatureTemplate.Type type = CreatureTemplate.Type.GreenLizard;
